backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
pydantic-settings>=2.6,<3.0
python-multipart>=0.0.9,<1.0
PyMuPDF>=1.25,<2.0
python-pptx>=1.0,<2.0
reportlab>=4.2,<5.0
openai>=1.54,<2.0
